requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
supabase==2.6.0
pydantic==2.8.2
httpx==0.27.0

